refactor(e2e-test): split monolithic e2e_test.rs into module per subcmd
Extract the single 1400-line e2e_test.rs into a multi-file binary crate
under src/bin/e2e-test/:

  callback.rs          — TestCallbackEvent, TestCallbackHandler,
                         TestCallbackServer (incl. last_queue_status)
  drr.rs               — cmd_drr, DrrArgs
  multi_file.rs        — cmd_multi_file, MultiFileArgs
  small_file_edge_cases.rs — cmd_small_file_edge_cases, SmallFileEdgeCasesArgs
  max_concurrent.rs    — cmd_max_concurrent, MaxConcurrentArgs
  main.rs              — CLI, Cli/Command, run_transfer, shared helpers
